+++ .tool-versions
@@ -1,1 +1,1 @@
-nodejs 20.9.0
+nodejs 20.10.0
